rpcv2: rename package serving to query
The sibling packages name what they are (ingest, lifecycle); this one is
the read side's machinery, whose noun everywhere else is query ("Query
Routing", query workers). query.ReadView reads as what it is, and the
rename is import paths only today — it grows a handler-file sweep the
moment #772's endpoints import it.

Addresses tamirms's review.
