+++ cmd/stellar-rpc/internal/rpcv2/query/registry_test.go
@@ -1,4 +1,4 @@
-package serving
+package query
 
 import (
 	"bytes"
