+++ cmd/stellar-rpc/internal/rpcv2/query/resolve_test.go
@@ -1,4 +1,4 @@
-package serving
+package query
 
 import (
 	"testing"
