+++ cmd/stellar-rpc/internal/rpcv2/query/scan.go
@@ -1,4 +1,4 @@
-package serving
+package query
 
 import (
 	"iter"
