+++ cmd/stellar-rpc/internal/rpcv2/query/scan_test.go
@@ -1,4 +1,4 @@
-package serving
+package query
 
 import (
 	"testing"
